requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
mysql-connector-python
pymysql
passlib[bcrypt]
python-jose[cryptography]